Username
fixed texts

+++ database/migrations/2016_02_01_221928_create_texts_table.php
@@ -13,8 +13,8 @@ class CreateTextsTable extends Migration
     public function up()
     {
         Schema::create('texts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
 
             $table->string('title')->index();
             $table->text('text');
